🔥 Update app and setting init
