--- backend/utils/dbscheme_import.py
@@ -1,2 +1,0 @@
-# TODO: function with sqlacodegen to generate models.py from
-#  the database schema and/or vis versa
